requirements.txt
opencv-contrib-python==3.4.2.16
kivy
pyserial
